extract transformData
